step4 refactoring about divide modulbe problem
