+++ module-api/src/main/java/kitchenpos/dto/ProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.dto;
+package kitchenpos.dto;
 
 import java.math.BigDecimal;
