+++ module-api/src/main/java/kitchenpos/ui/ProductRestController.java
@@ -1,8 +1,8 @@
-package kitchenpos.product.ui;
+package kitchenpos.ui;
 
-import kitchenpos.product.application.ProductService;
-import kitchenpos.product.dto.ProductRequest;
-import kitchenpos.product.dto.ProductResponse;
+import kitchenpos.dto.ProductResponse;
+import kitchenpos.application.ProductService;
+import kitchenpos.dto.ProductRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
