+++ module-api/src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.application;
+package kitchenpos.application;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
