--- module-domain/src/test/java/product/ProductTest.java
@@ -1,22 +1,0 @@
-package product;
-
-import kitchenpos.product.Product;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import java.math.BigDecimal;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-
-class ProductTest {
-
-    @Test
-    @DisplayName("상품 등록")
-    void register() {
-        Product product = new Product("강정치킨", new BigDecimal(-1));
-
-        assertThatThrownBy(() -> {
-            product.validationCheck();
-        }).isInstanceOf(IllegalArgumentException.class);
-    }
-}
